+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Closing edit view in backend shows empty page
+- Disabled table links are not marked in backend list view
 
 ## [3.0.0] - 2024-02-21